bring tools out, make more mcp-like
